Restore old pull request template
In commit 8424091 I tried to be clever and introduce the directory-style
pull request template, but that apparently only adds a named template
that you can access with a custom URL. The default template is still
.github/PULL_REQUEST_TEMPLATE.md, a file that didn't exist as of said
commit so we ended up without a template. Let's restore that file but
keep the directory-style issue template.